requirements/base.txt
Flask==3.0.2
flask-restx==1.3.0
pydantic==2.6.4
PyJWT==2.8.0
google-cloud-secret-manager==2.19.0
google-cloud-monitoring==2.19.3
google-cloud-datastore==2.19.0
gunicorn==21.2.0
bcrypt==4.1.2
email-validator==2.1.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
google-cloud-ndb==2.3.1
